feat: add  protected routes
