+++ src/components/Root.tsx
@@ -4,7 +4,7 @@ import Navigation from './Navigation';
 const Root = () => {
   return (
     <div className='h-full flex flex-col'>
-      <header className='p-4 flex justify-between items-center border-b-2 border-blue-300'>
+      <header className='p-4 flex justify-between items-center border-b border-blue-300'>
         <div className='w-12'>
           <Link to='/'>
             <img
@@ -19,7 +19,9 @@ const Root = () => {
       <main className='grow shrink-0'>
         <Outlet />
       </main>
-      <footer className='p-4'>Footer</footer>
+      <footer className='p-4 border-t border-green-300 text-center'>
+        {new Date().getFullYear()}
+      </footer>
     </div>
   );
 };
