+++ src/components/auth/FormSignIn.tsx
@@ -1,15 +1,14 @@
 import { useForm, SubmitHandler } from 'react-hook-form';
-import { EMAIL_REGEX } from '../constants.ts';
-import { auth } from '../firebase.ts';
-import { createUserWithEmailAndPassword, updateProfile } from 'firebase/auth';
+import { EMAIL_REGEX } from '../../constants.ts';
+import { auth } from '../../firebase.ts';
+import { signInWithEmailAndPassword } from 'firebase/auth';
 
 type Inputs = {
-  username?: string;
   email: string;
   password: string;
 };
 
-const FormSignUp = () => {
+const FormSignIn = () => {
   const {
     register,
     handleSubmit,
@@ -17,17 +16,10 @@ const FormSignUp = () => {
     formState: { errors },
   } = useForm<Inputs>();
 
-  const signUp: SubmitHandler<Inputs> = (data) => {
-    console.log(data);
-    createUserWithEmailAndPassword(auth, data.email, data.password)
+  const signIn: SubmitHandler<Inputs> = (data) => {
+    signInWithEmailAndPassword(auth, data.email, data.password)
       .then((userCredential) => {
         console.log(userCredential);
-        return updateProfile(userCredential.user, {
-          displayName: data.username,
-        });
-      })
-      .then(() => {
-        console.log('User profile updated successfully');
       })
       .catch((err) => console.log(err));
   };
@@ -35,27 +27,11 @@ const FormSignUp = () => {
   return (
     <div className='w-full h-full flex justify-center items-center '>
       <form
-        onSubmit={handleSubmit(signUp)}
-        className='p-8 w-80 flex flex-col border border-slate-200 shadow-lg rounded'
+        onSubmit={handleSubmit(signIn)}
+        className='p-8 w-80 flex flex-col border border-slate-200 shadow-lg rounded-sm'
       >
-        <h1 className='mb-8 self-center text-xl'>Регистрация</h1>
-        <div className='flex flex-col mb-2 '>
-          <input
-            className='hover:border-green-100 focus:border-green-300 transition-colors'
-            type='text'
-            placeholder='Имя'
-            defaultValue=''
-            {...register('username', {
-              minLength: { value: 6, message: 'Минимум 3 символа' },
-            })}
-          />
-          {
-            <span className=' h-5 px-1 text-red-400 text-right text-xs'>
-              {errors.username && errors.username.message}
-            </span>
-          }
-        </div>
-
+        <h1 className='mb-8 self-center text-xl'>Вход</h1>
+        {/* TODO: refactor to separate component Input */}
         <div className='flex flex-col mb-2 '>
           <input
             className='hover:border-green-100 focus:border-green-300 transition-colors'
@@ -75,7 +51,6 @@ const FormSignUp = () => {
             </span>
           }
         </div>
-
         <div className='flex flex-col'>
           <input
             className='hover:border-green-100 focus:border-green-300 transition-colors'
@@ -92,15 +67,14 @@ const FormSignUp = () => {
             </span>
           }
         </div>
-
         <button
           className='px-4 py-2 self-center bg-green-400 rounded mt-4 hover:bg-opacity-80 transition-colors'
           type='submit'
         >
-          Создать аккаунт
+          Войти
         </button>
       </form>
     </div>
   );
 };
-export default FormSignUp;
+export default FormSignIn;
