+++ src/components/pages/ProfilePage.tsx
@@ -1,0 +1,4 @@
+const ProfilePage = () => {
+  return <div>ProfilePage</div>;
+};
+export default ProfilePage;
